Bump recovery code exhaustion test timeout for slow runners, consolidate per-OS testutil files into platform_unix/windows
