--- tests/integration/testutil/ipc_unix.go
@@ -1,55 +1,0 @@
-/*
-Copyright NetFoundry Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-//go:build !windows
-
-package testutil
-
-import (
-	"fmt"
-	"net"
-	"os"
-	"time"
-)
-
-func RequireAdmin() error {
-	if os.Geteuid() == 0 {
-		return nil
-	}
-	return fmt.Errorf("integration tests must run as root; rerun under sudo")
-}
-
-const CommandPipePath = "/tmp/.ziti/ziti-edge-tunnel.sock"
-const EventPipePath = "/tmp/.ziti/ziti-edge-tunnel-event.sock"
-
-func CommandPipePathFor(disc string) string {
-	if disc == "" {
-		return CommandPipePath
-	}
-	return CommandPipePath + "." + disc
-}
-
-func EventPipePathFor(disc string) string {
-	if disc == "" {
-		return EventPipePath
-	}
-	return EventPipePath + "." + disc
-}
-
-func dialPlatform(path string, timeout time.Duration) (net.Conn, error) {
-	d := net.Dialer{Timeout: timeout}
-	return d.Dial("unix", path)
-}
