--- tests/integration/testutil/process_unix.go
@@ -1,32 +1,0 @@
-//go:build !windows
-
-/*
-Copyright NetFoundry Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package testutil
-
-import (
-	"os/exec"
-	"syscall"
-)
-
-// relayStop sends SIGINT, which `quickstart cluster` handles by shutting its node
-// children down cleanly.
-func relayStop(cmd *exec.Cmd) {
-	if cmd.Process != nil {
-		_ = cmd.Process.Signal(syscall.SIGINT)
-	}
-}
